Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 21ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240515/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -15,7 +15,7 @@ int main() {
         int cnt = 0;
         while (l <= r) {
             cnt++;
-            int mid = (l + r) / 2;
+            ll mid = (l + r) / 2;
             if (mid == i) {
                 min_cnt = min(min_cnt, cnt);
                 max_cnt = max(max_cnt, cnt);
